Refactor C4D publish plugins to align with AYON structure
- Consolidated context collection into `collect_context.py` (merging `collect_current_file.py` and `collect_current_document.py`).
- Added `validate_workfile.py` to explicitly validate if the workfile is saved.
- Renamed `save_scene.py` to `extract_save_scene.py` to better reflect its phase.
- Renamed `increment_current_file.py` to `increment_workfile.py` to match standard naming.
- Updated class names and labels for consistency.
- Deleted obsolete files.
